# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Color Matcher API",
    "description": "Color analysis and paint-mix hints. Cameras and ROI live in the browser; the server only receives RGB values via query parameters or JSON body.",
    "version": "0.3.0"
  },
  "paths": {
    "/api/health": {
      "get": {
        "tags": [
          "health"
        ],
        "summary": "Health check",
        "operationId": "health_api_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/config": {
      "get": {
        "tags": [
          "health"
        ],
        "summary": "UI defaults",
        "description": "Default ROI size and reference frame dimensions (cameras run in the browser).",
        "operationId": "get_config_api_config_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/palette-presets": {
      "get": {
        "tags": [
          "palettes"
        ],
        "summary": "List server palette presets",
        "description": "Tagged base-color sets defined in code (count / manufacturer).",
        "operationId": "list_palette_presets_api_palette_presets_get",
        "parameters": [
          {
            "name": "Accept-Language",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Accept-Language"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PalettePresetsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/palette-presets/{preset_id}": {
      "get": {
        "tags": [
          "palettes"
        ],
        "summary": "Get preset colors",
        "operationId": "get_palette_preset_api_palette_presets__preset_id__get",
        "parameters": [
          {
            "name": "preset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Preset Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PalettePresetColorsResponse"
                }
              }
            }
          },
          "404": {
            "description": "Unknown preset id"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/base-colors": {
      "get": {
        "tags": [
          "colors"
        ],
        "summary": "Default base paint catalog",
        "description": "Legacy alias for the default server preset (`classic-10`).",
        "operationId": "get_base_colors_api_base_colors_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseColorsResponse"
                }
              }
            }
          }
        },
        "deprecated": true
      }
    },
    "/api/analyze-rgb": {
      "post": {
        "tags": [
          "colors"
        ],
        "summary": "Analyze RGB",
        "description": "Convert a single RGB sample (from canvas/ROI) to Lab, HSL, CMYK, and hex.",
        "operationId": "analyze_rgb_endpoint_api_analyze_rgb_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RgbBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ColorReadingResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/match": {
      "post": {
        "tags": [
          "match"
        ],
        "summary": "Match with client base colors",
        "description": "Compare target vs palette using the active base paint set from the browser.\n\nUser-edited palettes stay on the client; only RGB triplets are sent here.",
        "operationId": "post_match_api_match_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MatchRequestBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "match"
        ],
        "summary": "Match target vs palette (default server preset)",
        "description": "Compare two browser-sampled colors using the default server preset.\n\nPrefer `POST /api/match` with `baseColors` when the user edits palettes locally.",
        "operationId": "get_match_api_match_get",
        "parameters": [
          {
            "name": "targetR",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 255,
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "description": "Target color red (0–255). Required with targetG and targetB.",
              "examples": [
                200
              ],
              "title": "Targetr"
            },
            "description": "Target color red (0–255). Required with targetG and targetB."
          },
          {
            "name": "targetG",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 255,
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "description": "Target color green (0–255).",
              "examples": [
                120
              ],
              "title": "Targetg"
            },
            "description": "Target color green (0–255)."
          },
          {
            "name": "targetB",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 255,
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "description": "Target color blue (0–255).",
              "examples": [
                80
              ],
              "title": "Targetb"
            },
            "description": "Target color blue (0–255)."
          },
          {
            "name": "paletteR",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 255,
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "description": "Palette color red (0–255). Required with paletteG and paletteB.",
              "examples": [
                180
              ],
              "title": "Paletter"
            },
            "description": "Palette color red (0–255). Required with paletteG and paletteB."
          },
          {
            "name": "paletteG",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 255,
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "description": "Palette color green (0–255).",
              "examples": [
                100
              ],
              "title": "Paletteg"
            },
            "description": "Palette color green (0–255)."
          },
          {
            "name": "paletteB",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 255,
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "description": "Palette color blue (0–255).",
              "examples": [
                70
              ],
              "title": "Paletteb"
            },
            "description": "Palette color blue (0–255)."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchResponse"
                }
              }
            }
          },
          "400": {
            "description": "Missing or incomplete RGB query parameters"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "BaseColorEntry": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "examples": [
              "Cadmium Red Medium"
            ]
          },
          "rgb": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "maxItems": 3,
            "minItems": 3,
            "title": "Rgb",
            "examples": [
              [
                227,
                38,
                54
              ]
            ]
          }
        },
        "type": "object",
        "required": [
          "name",
          "rgb"
        ],
        "title": "BaseColorEntry"
      },
      "BaseColorsResponse": {
        "properties": {
          "colors": {
            "items": {
              "$ref": "#/components/schemas/BaseColorEntry"
            },
            "type": "array",
            "title": "Colors"
          }
        },
        "type": "object",
        "required": [
          "colors"
        ],
        "title": "BaseColorsResponse"
      },
      "CmykChannels": {
        "properties": {
          "c": {
            "type": "number",
            "title": "C"
          },
          "m": {
            "type": "number",
            "title": "M"
          },
          "y": {
            "type": "number",
            "title": "Y"
          },
          "k": {
            "type": "number",
            "title": "K"
          }
        },
        "type": "object",
        "required": [
          "c",
          "m",
          "y",
          "k"
        ],
        "title": "CmykChannels"
      },
      "ColorReadingResponse": {
        "properties": {
          "rgb": {
            "$ref": "#/components/schemas/RgbChannels"
          },
          "hex": {
            "type": "string",
            "title": "Hex",
            "examples": [
              "#e32636"
            ]
          },
          "hsl": {
            "$ref": "#/components/schemas/HslChannels"
          },
          "lab": {
            "$ref": "#/components/schemas/LabChannels"
          },
          "cmyk": {
            "$ref": "#/components/schemas/CmykChannels"
          }
        },
        "type": "object",
        "required": [
          "rgb",
          "hex",
          "hsl",
          "lab",
          "cmyk"
        ],
        "title": "ColorReadingResponse"
      },
      "ConfigResponse": {
        "properties": {
          "roiSize": {
            "type": "integer",
            "title": "Roisize",
            "examples": [
              80
            ]
          },
          "frameWidth": {
            "type": "integer",
            "title": "Framewidth",
            "examples": [
              640
            ]
          },
          "frameHeight": {
            "type": "integer",
            "title": "Frameheight",
            "examples": [
              480
            ]
          }
        },
        "type": "object",
        "required": [
          "roiSize",
          "frameWidth",
          "frameHeight"
        ],
        "title": "ConfigResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HealthResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status",
            "examples": [
              "ok"
            ]
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "HealthResponse"
      },
      "HslChannels": {
        "properties": {
          "h": {
            "type": "number",
            "title": "H"
          },
          "s": {
            "type": "number",
            "title": "S"
          },
          "l": {
            "type": "number",
            "title": "L"
          }
        },
        "type": "object",
        "required": [
          "h",
          "s",
          "l"
        ],
        "title": "HslChannels"
      },
      "LabChannels": {
        "properties": {
          "l": {
            "type": "number",
            "title": "L"
          },
          "a": {
            "type": "number",
            "title": "A"
          },
          "b": {
            "type": "number",
            "title": "B"
          }
        },
        "type": "object",
        "required": [
          "l",
          "a",
          "b"
        ],
        "title": "LabChannels"
      },
      "MatchRequestBody": {
        "properties": {
          "target": {
            "$ref": "#/components/schemas/RgbBody"
          },
          "palette": {
            "$ref": "#/components/schemas/RgbBody"
          },
          "baseColors": {
            "items": {
              "$ref": "#/components/schemas/BaseColorEntry"
            },
            "type": "array",
            "title": "Basecolors",
            "description": "Active base paint set from the browser (client-only storage)."
          }
        },
        "type": "object",
        "required": [
          "target",
          "palette"
        ],
        "title": "MatchRequestBody"
      },
      "MatchResponse": {
        "properties": {
          "deltaE": {
            "type": "number",
            "title": "Deltae",
            "examples": [
              12.4
            ]
          },
          "target": {
            "$ref": "#/components/schemas/ColorReadingResponse"
          },
          "palette": {
            "$ref": "#/components/schemas/ColorReadingResponse"
          },
          "targetCaptured": {
            "type": "boolean",
            "title": "Targetcaptured",
            "examples": [
              true
            ]
          },
          "paletteCaptured": {
            "type": "boolean",
            "title": "Palettecaptured",
            "examples": [
              true
            ]
          },
          "mix": {
            "$ref": "#/components/schemas/MixSuggestion"
          }
        },
        "type": "object",
        "required": [
          "deltaE",
          "target",
          "palette",
          "targetCaptured",
          "paletteCaptured",
          "mix"
        ],
        "title": "MatchResponse"
      },
      "MixComponent": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "rgb": {
            "$ref": "#/components/schemas/RgbChannels"
          },
          "weight": {
            "type": "number",
            "title": "Weight"
          },
          "deltaE_to_target": {
            "type": "number",
            "title": "Deltae To Target"
          },
          "percent": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Percent"
          }
        },
        "type": "object",
        "required": [
          "name",
          "rgb",
          "weight",
          "deltaE_to_target"
        ],
        "title": "MixComponent"
      },
      "MixSuggestion": {
        "properties": {
          "available": {
            "type": "boolean",
            "title": "Available"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "deltaE_current_to_target": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deltae Current To Target"
          },
          "components": {
            "items": {
              "$ref": "#/components/schemas/MixComponent"
            },
            "type": "array",
            "title": "Components"
          },
          "legend": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Legend"
          }
        },
        "type": "object",
        "required": [
          "available"
        ],
        "title": "MixSuggestion"
      },
      "PalettePresetColorsResponse": {
        "properties": {
          "presetId": {
            "type": "string",
            "title": "Presetid"
          },
          "colors": {
            "items": {
              "$ref": "#/components/schemas/BaseColorEntry"
            },
            "type": "array",
            "title": "Colors"
          }
        },
        "type": "object",
        "required": [
          "presetId",
          "colors"
        ],
        "title": "PalettePresetColorsResponse"
      },
      "PalettePresetSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "tag": {
            "type": "string",
            "title": "Tag"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "colorCount": {
            "type": "integer",
            "title": "Colorcount"
          }
        },
        "type": "object",
        "required": [
          "id",
          "tag",
          "category",
          "label",
          "description",
          "colorCount"
        ],
        "title": "PalettePresetSummary"
      },
      "PalettePresetsResponse": {
        "properties": {
          "defaultPresetId": {
            "type": "string",
            "title": "Defaultpresetid"
          },
          "presets": {
            "items": {
              "$ref": "#/components/schemas/PalettePresetSummary"
            },
            "type": "array",
            "title": "Presets"
          }
        },
        "type": "object",
        "required": [
          "defaultPresetId",
          "presets"
        ],
        "title": "PalettePresetsResponse"
      },
      "RgbBody": {
        "properties": {
          "r": {
            "type": "integer",
            "maximum": 255.0,
            "minimum": 0.0,
            "title": "R",
            "examples": [
              227
            ]
          },
          "g": {
            "type": "integer",
            "maximum": 255.0,
            "minimum": 0.0,
            "title": "G",
            "examples": [
              38
            ]
          },
          "b": {
            "type": "integer",
            "maximum": 255.0,
            "minimum": 0.0,
            "title": "B",
            "examples": [
              54
            ]
          }
        },
        "type": "object",
        "required": [
          "r",
          "g",
          "b"
        ],
        "title": "RgbBody",
        "description": "RGB triplet sampled in the browser (0–255 per channel)."
      },
      "RgbChannels": {
        "properties": {
          "r": {
            "type": "integer",
            "maximum": 255.0,
            "minimum": 0.0,
            "title": "R",
            "examples": [
              200
            ]
          },
          "g": {
            "type": "integer",
            "maximum": 255.0,
            "minimum": 0.0,
            "title": "G",
            "examples": [
              120
            ]
          },
          "b": {
            "type": "integer",
            "maximum": 255.0,
            "minimum": 0.0,
            "title": "B",
            "examples": [
              80
            ]
          }
        },
        "type": "object",
        "required": [
          "r",
          "g",
          "b"
        ],
        "title": "RgbChannels"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
  "tags": [
    {
      "name": "health",
      "description": "Liveness and UI defaults."
    },
    {
      "name": "colors",
      "description": "RGB → Lab/HSL/CMYK and base paint catalog."
    },
    {
      "name": "palettes",
      "description": "Server-defined base color presets (read-only, in code)."
    },
    {
      "name": "match",
      "description": "ΔE and mix suggestion from browser-sampled RGB pairs."
    }
  ]
}